Added Withdrawal Function
we again use the modifiers created earlier but this time we wantwithdrawalDeadlineReached() to be true and claimDeadlineReached() to be false.

This set of modifiers/parameters means that we are in the sweet spot for the withdrawal window since its time for the withdrawal to take place without any penalties and we get interest as well.

The rest of the function does a few important steps.

    It checks to ensure that the person trying to withdraw ETH actually has a non-zero stake.
    It calculates the amount of ETH owed in interest by taking the number of blocks that passed from deposit to withdrawal and multiplying that by our interest constant.
    It sets the user's balance staked ETH to 0 so that no double counting can occur.
    It transfers the ETH from the smart contract back to the user's wallet.